Fix preload import
