+++ apps/studio/electron/preload/webview/elements/dom/image.ts
@@ -1,4 +1,4 @@
-import { StyleChangeType } from '@onlook/models/actions';
+import { StyleChangeType } from '@onlook/models/style';
 import cssManager from '../../style';
 
 export function insertImage(domId: string, image: string) {
