+++ apps/studio/src/lib/editor/engine/code/insert.ts
@@ -1,6 +1,7 @@
 import type { ActionElement, ActionLocation, PasteParams } from '@onlook/models/actions';
-import { CodeActionType, StyleChangeType, type CodeInsert } from '@onlook/models/actions';
+import { CodeActionType, type CodeInsert } from '@onlook/models/actions';
 import { EditorAttributes } from '@onlook/models/constants';
+import { StyleChangeType } from '@onlook/models/style';
 import { twMerge } from 'tailwind-merge';
 import { getTailwindClasses } from './helpers';
