+++ apps/studio/src/lib/editor/engine/insert/index.ts
@@ -2,7 +2,6 @@ import { EditorMode } from '@/lib/models';
 import { createDomId, createOid } from '@/lib/utils';
 import type { ImageContentData } from '@onlook/models';
 import {
-    StyleChangeType,
     type ActionElement,
     type ActionLocation,
     type ActionTarget,
@@ -11,6 +10,7 @@ import {
 } from '@onlook/models/actions';
 import { DefaultSettings, EditorAttributes } from '@onlook/models/constants';
 import type { DropElementProperties, ElementPosition } from '@onlook/models/element';
+import { StyleChangeType } from '@onlook/models/style';
 import { colors } from '@onlook/ui/tokens';
 import type React from 'react';
 import type { EditorEngine } from '..';
