+++ packages/models/src/actions/target.ts
@@ -1,3 +1,5 @@
+import type { StyleChange } from '../style';
+
 export type Change<T> = {
     updated: T;
     original: T;
@@ -7,17 +9,6 @@ export interface ActionTarget {
     domId: string;
     oid: string | null;
     webviewId: string;
-}
-
-export interface StyleChange {
-    value: string;
-    type: StyleChangeType;
-}
-
-export enum StyleChangeType {
-    Value = 'value',
-    Custom = 'custom',
-    Remove = 'remove',
 }
 
 export interface StyleActionTarget extends ActionTarget {
